Favicon data saves to app.locals, replaced uploading favicon to app.ts
