+++ src/backend/app.ts
@@ -8,16 +8,31 @@ import HttpException from './exceptions/httpException';
 import * as dotenv from 'dotenv';
 import config from 'config';
 import os from 'os';
+import appConfig from 'config';
+import { uploadFavicon } from './utils/uploadFavicon';
 
 dotenv.config();
 const app = express();
 const localConfig = rcParser.getConfiguration();
+
+// Get url to upload favicon from config
+const faviconURL: string = appConfig.get('faviconURL');
 
 app.locals.config = localConfig;
 // view engine setup
 app.set('views', path.join(__dirname, './', 'views'));
 app.set('view engine', 'twig');
 require('./utils/twig');
+
+// Upload favicon by url, it's path on server is '/temp/favicon.{format}'
+uploadFavicon(faviconURL).then((res) => {
+  app.locals.favicon = res;
+  console.log('Favicon successfully uploaded');
+})
+  .catch( (err) => {
+    console.log(err);
+    console.log('Favicon has not uploaded');
+  });
 
 app.use(morgan('dev'));
 app.use(express.json());
