+++ src/backend/routes/auth.ts
@@ -2,6 +2,7 @@ import express, { Request, Response } from 'express';
 import jwt from 'jsonwebtoken';
 import config from 'config';
 import csrf from 'csurf';
+import app from '../app';
 
 const router = express.Router();
 const csrfProtection = csrf({ cookie: true });
@@ -14,6 +15,7 @@ router.get('/auth', csrfProtection, function (req: Request, res: Response) {
   res.render('auth', {
     title: 'Login page',
     csrfToken: req.csrfToken(),
+    favicon: app.locals.favicon,
   });
 });
 
@@ -27,6 +29,7 @@ router.post('/auth', parseForm, csrfProtection, async (req: Request, res: Respon
         title: 'Login page',
         header: 'Password not set',
         csrfToken: req.csrfToken(),
+        favicon: app.locals.favicon,
       });
 
       return;
@@ -37,6 +40,7 @@ router.post('/auth', parseForm, csrfProtection, async (req: Request, res: Respon
         title: 'Login page',
         header: 'Wrong password',
         csrfToken: req.csrfToken(),
+        favicon: app.locals.favicon,
       });
 
       return;
@@ -59,6 +63,7 @@ router.post('/auth', parseForm, csrfProtection, async (req: Request, res: Respon
       title: 'Login page',
       header: 'Password not set',
       csrfToken: req.csrfToken(),
+      favicon: app.locals.favicon
     });
 
     return;
