+++ src/backend/routes/pages.ts
@@ -3,6 +3,7 @@ import Pages from '../controllers/pages';
 import PagesOrder from '../controllers/pagesOrder';
 import verifyToken from './middlewares/token';
 import allowEdit from './middlewares/locals';
+import app from '../app';
 
 const router = express.Router();
 
@@ -16,6 +17,7 @@ router.get('/page/new', verifyToken, allowEdit, async (req: Request, res: Respon
     res.render('pages/form', {
       pagesAvailable,
       page: null,
+      favicon: app.locals.favicon,
     });
   } catch (error) {
     res.status(404);
@@ -43,6 +45,7 @@ router.get('/page/edit/:id', verifyToken, allowEdit, async (req: Request, res: R
       page,
       parentsChildrenOrdered,
       pagesAvailable,
+      favicon: app.locals.favicon,
     });
   } catch (error) {
     res.status(404);
@@ -65,6 +68,7 @@ router.get('/page/:id', verifyToken, async (req: Request, res: Response, next: N
       page,
       pageParent,
       config: req.app.locals.config,
+      favicon: app.locals.favicon,
     });
   } catch (error) {
     res.status(404);
