+++ src/backend/utils/uploadFavicon.ts
@@ -2,14 +2,17 @@ import path from 'path';
 import os from 'os';
 import fs from 'fs';
 import fetch from 'node-fetch';
-import config from 'config';
 
+/**
+ * Uploaded favicon data
+ */
 interface FaviconData {
+  // Uploaded favicon path
   destination: string;
+
+  // File type
   type: string;
 }
-
-const favicon = config.get<FaviconData>('favicon');
 
 /**
  * Check if string is url
@@ -25,24 +28,30 @@ function checkIsUrl(str: string): boolean {
 /**
  * Upload favicon by url
  *
- * @param url - url for uploading favicon
+ * @param destination - url for uploading favicon
  * @returns { Promise<string> } - Promise with format of saved file
  */
-async function uploadFavicon(url: string): Promise<string> {
-  // Check if string is url
-  if (!checkIsUrl(url)) {
-    return url;
+export async function uploadFavicon(destination: string): Promise<FaviconData> {
+  // Check of destination is empty
+  if (!destination) {
+    throw Error('Favicon destination is empty');
   }
-  // Make get request to url
-  const res = await fetch(url);
-  // Get buffer data from response
-  const fileData = await res.buffer();
 
-  // Get file name by url
-  const filename = url.substring(url.lastIndexOf('/')+1);
+  // Get file name by destination
+  const filename = destination.substring(destination.lastIndexOf('/')+1);
 
   // Get file format
   const format = filename.split('.')[1];
+
+  // Check if string is url
+  if (!checkIsUrl(destination)) {
+    return  { destination: destination,
+      type: `image/${format}` } as FaviconData;
+  }
+  // Make get request to url
+  const res = await fetch(destination);
+  // Get buffer data from response
+  const fileData = await res.buffer();
 
   // Get file path in temporary directory
   const filePath = path.join(os.tmpdir(), `favicon.${format}`);
@@ -50,5 +59,6 @@ async function uploadFavicon(url: string): Promise<string> {
   // Save file
   fs.writeFileSync(filePath, fileData);
 
-  return `/favicon/favicon.${format}`;
+  return  { destination: `/favicon/favicon.${format}`,
+    type: `image/${format}` } as FaviconData;
 }
